+++ sdks/python/hatchet_sdk/types/trigger.py
@@ -24,6 +24,7 @@ class TriggerWorkflowOptions(ScheduleTriggerWorkflowOptions):
     desired_worker_label: (
         dict[str, DesiredWorkerLabel] | list[DesiredWorkerLabel] | None
     ) = None
+    display_name: str | None = None
 
 
 class WorkflowRunTriggerConfig(BaseModel):